+++ library/CHANGELOG.md
@@ -6,7 +6,8 @@ All notable changes to the library will be documented in this file.
 
 - Moven origin property of `SchemaIssue` type into `PathItem`
 - Rename `validation` property of `Issue` type to `context`
-- Remove length check in `tuple` and `tupleAsync` schema to fix issue info
+- Fix issue info in `tuple` and `tupleAsync` schema by removing length check
+- Fix issue info for discriminator key in `variant` and `variantAsync` schema
 
 ## v0.27.1 (January 28, 2023)